[BACKEND]-Mechanizm przypisywania zgłoszeń

+++ pjaidApp/backend/pjaidJavaApp/src/main/java/org/api/pjaidapp/repository/TicketRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import java.util.Collection;
 import java.util.List;
 
-public interface TicketRepository extends JpaRepository<Ticket, Integer>, JpaSpecificationExecutor<Ticket> {
+public interface TicketRepository extends JpaRepository<Ticket, Long>, JpaSpecificationExecutor<Ticket> {
     List<Ticket> findByStatusIn(Collection<Status> status);
 }
